complete

+++ src/main/java/com/example/todo_app/repository/UserRepository.java
@@ -9,12 +9,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 @Repository
-public interface UserRepository extends JpaRepository<User,Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findUserByEmail(String email);
-
-//    @Transactional
-//    @Modifying
-//    @Query("UPDATE User a " +
-//            "SET a.enabled = TRUE WHERE a.email = ?1")
-//    int enableUser(String email);
 }
